requirements.txt
mysql-connector-python
python-dotenv
sqlparse
langchain
requests
transformers
tiktoken
pandas
